refactor: refactor agent management and enhance project creation with actor user ID support
- Removed unused imports and components from the agents index page.
- Simplified the AgentCard component by eliminating unnecessary logic and UI elements.
- Updated the build_mcp_config function to accept an optional actor_user_id parameter for better tracking of actions performed on behalf of users.
- Modified the run_conversation function to pass the actor_user_id when building the MCP configuration.
- Added tests to ensure the correct inclusion and omission of the actor_user_id in the MCP server configuration.
- Enhanced the project handler to attribute project creation to the correct user when an actor user ID is provided.
- Implemented middleware changes to parse and store the actor user ID from requests, ensuring proper context handling for agent-authenticated requests.
