+++ apps/mcp/src/index.ts
@@ -25,6 +25,7 @@ async function main() {
 	const gatewayURL = process.env.PACA_GATEWAY_URL || undefined;
 	const agentId = process.env.PACA_AGENT_ID || undefined;
 	const projectId = process.env.PACA_PROJECT_ID || undefined;
+	const actorUserId = process.env.PACA_ACTOR_USER_ID || undefined;
 
 	// Validate required configuration
 	if (!apiKey) {
@@ -54,6 +55,7 @@ async function main() {
 		gatewayURL,
 		agentId,
 		projectId,
+		actorUserId,
 	};
 
 	// Create and configure MCP server (loads plugin modules asynchronously)
