+++ apps/web/src/i18n/locales/fr/projects.json
@@ -76,6 +76,7 @@
 			"deleteDialog": {
 				"title": "Supprimer {{name}} ?",
 				"description": "Cette action supprime définitivement l'agent et le retire du projet. Les conversations en cours seront arrêtées.",
+				"descriptionGlobal": "Cette action supprime définitivement l'agent et le retire de tous les projets où il a été invité. Les conversations en cours seront arrêtées.",
 				"cancel": "Annuler",
 				"delete": "Supprimer"
 			}
