+++ apps/web/src/i18n/locales/ja/projects.json
@@ -76,6 +76,7 @@
 			"deleteDialog": {
 				"title": "{{name}}を削除しますか？",
 				"description": "このエージェントを完全に削除し、プロジェクトから除外します。実行中の会話は停止されます。",
+				"descriptionGlobal": "このエージェントを完全に削除し、招待されているすべてのプロジェクトから除外します。実行中の会話は停止されます。",
 				"cancel": "キャンセル",
 				"delete": "削除"
 			}
