+++ apps/web/src/i18n/locales/ko/admin.json
@@ -279,39 +279,6 @@
 			"title": "아직 전역 에이전트가 없습니다",
 			"description": "전역 에이전트를 만들면 채팅과 프로젝트 초대에 사용할 수 있습니다.",
 			"createAgent": "에이전트 생성"
-		},
-		"table": {
-			"editAction": "에이전트 수정",
-			"deleteAction": "에이전트 삭제"
-		},
-		"formDialog": {
-			"editTitle": "전역 에이전트 수정",
-			"description": "전역 에이전트는 자체 프로젝트가 없습니다 — 해당 프로젝트의 팀 페이지에서 초대하여 그곳에서 작업하게 할 수 있습니다.",
-			"nameLabel": "이름",
-			"handleLabel": "핸들",
-			"globalRoleLabel": "전역 역할",
-			"noGlobalRole": "역할 없음",
-			"globalRoleHint": "이 에이전트가 전역 범위(예: 홈/관리 채팅)에서 할 수 있는 작업을 결정합니다. 필수는 아닙니다 — 프로젝트에 초대된 후 부여되는 역할과는 별개입니다.",
-			"acpProviderLabel": "ACP 제공자",
-			"providerLabel": "제공자",
-			"modelLabel": "모델",
-			"customOption": "사용자 지정…",
-			"apiKeyLabel": "API 키",
-			"apiKeyKeepPlaceholder": "현재 키를 유지하려면 비워두세요",
-			"baseUrlLabel": "기본 URL",
-			"systemPromptLabel": "시스템 프롬프트",
-			"optional": "선택 사항",
-			"cancel": "취소",
-			"save": "변경사항 저장",
-			"updateFailed": "에이전트 업데이트에 실패했습니다"
-		},
-		"deleteDialog": {
-			"title": "에이전트를 삭제하시겠습니까?",
-			"confirmPrefix": "이 작업은",
-			"confirmSuffix": "초대된 모든 프로젝트에서 제거됩니다. 이 작업은 되돌릴 수 없습니다.",
-			"cancel": "취소",
-			"delete": "삭제",
-			"deleteFailed": "에이전트 삭제에 실패했습니다"
 		}
 	}
 }
