+++ apps/web/src/i18n/locales/ko/projects.json
@@ -76,6 +76,7 @@
 			"deleteDialog": {
 				"title": "{{name}}을(를) 삭제하시겠습니까?",
 				"description": "이 작업은 에이전트를 영구적으로 삭제하고 프로젝트에서 제거합니다. 진행 중인 대화는 중단됩니다.",
+				"descriptionGlobal": "이 작업은 에이전트를 영구적으로 삭제하고 초대된 모든 프로젝트에서 제거합니다. 진행 중인 대화는 중단됩니다.",
 				"cancel": "취소",
 				"delete": "삭제"
 			}
