+++ apps/web/src/i18n/locales/pt-BR/projects.json
@@ -76,6 +76,7 @@
 			"deleteDialog": {
 				"title": "Excluir {{name}}?",
 				"description": "Isso exclui permanentemente o agente e o remove do projeto. As conversas em andamento serão interrompidas.",
+				"descriptionGlobal": "Isso exclui permanentemente o agente e o remove de todos os projetos aos quais foi convidado. As conversas em andamento serão interrompidas.",
 				"cancel": "Cancelar",
 				"delete": "Excluir"
 			}
