+++ apps/web/src/i18n/locales/ru/admin.json
@@ -285,39 +285,6 @@
 			"title": "Глобальных агентов пока нет",
 			"description": "Создайте глобального агента, чтобы сделать его доступным для чата и приглашений в проекты.",
 			"createAgent": "Создать агента"
-		},
-		"table": {
-			"editAction": "Изменить агента",
-			"deleteAction": "Удалить агента"
-		},
-		"formDialog": {
-			"editTitle": "Изменить глобального агента",
-			"description": "У глобальных агентов нет собственного проекта — пригласите его в проект со страницы «Команда» этого проекта, чтобы он начал там работать.",
-			"nameLabel": "Название",
-			"handleLabel": "Handle",
-			"globalRoleLabel": "Глобальная роль",
-			"noGlobalRole": "Без роли",
-			"globalRoleHint": "Определяет, что этот агент может делать в глобальной области (например, в чате на главной странице или в админке). Не обязательна — роль, назначенная позже, после приглашения в проект, отдельна.",
-			"acpProviderLabel": "Провайдер ACP",
-			"providerLabel": "Провайдер",
-			"modelLabel": "Модель",
-			"customOption": "Другая…",
-			"apiKeyLabel": "API-ключ",
-			"apiKeyKeepPlaceholder": "Оставьте пустым, чтобы сохранить текущий ключ",
-			"baseUrlLabel": "Base URL",
-			"systemPromptLabel": "Системный промпт",
-			"optional": "необязательно",
-			"cancel": "Отмена",
-			"save": "Сохранить изменения",
-			"updateFailed": "Не удалось обновить агента"
-		},
-		"deleteDialog": {
-			"title": "Удалить агента?",
-			"confirmPrefix": "Это навсегда удалит",
-			"confirmSuffix": ", удалив его из всех проектов, в которые он был приглашён. Это действие нельзя отменить.",
-			"cancel": "Отмена",
-			"delete": "Удалить",
-			"deleteFailed": "Не удалось удалить агента"
 		}
 	}
 }
