+++ apps/web/src/i18n/locales/ru/projects.json
@@ -76,6 +76,7 @@
 			"deleteDialog": {
 				"title": "Удалить {{name}}?",
 				"description": "Агент будет удалён навсегда и убран из проекта. Запущенные диалоги будут остановлены.",
+				"descriptionGlobal": "Агент будет удалён навсегда и убран из всех проектов, в которые он был приглашён. Запущенные диалоги будут остановлены.",
 				"cancel": "Отмена",
 				"delete": "Удалить"
 			}
