+++ apps/web/src/i18n/locales/vi/admin.json
@@ -279,39 +279,6 @@
 			"title": "Chưa có agent toàn cục nào",
 			"description": "Tạo một agent toàn cục để có thể trò chuyện và mời vào dự án.",
 			"createAgent": "Tạo agent"
-		},
-		"table": {
-			"editAction": "Chỉnh sửa agent",
-			"deleteAction": "Xóa agent"
-		},
-		"formDialog": {
-			"editTitle": "Chỉnh sửa Agent toàn cục",
-			"description": "Agent toàn cục không có dự án riêng — hãy mời agent vào một dự án từ trang Nhóm của dự án đó để agent làm việc ở đó.",
-			"nameLabel": "Tên",
-			"handleLabel": "Handle",
-			"globalRoleLabel": "Vai trò toàn cục",
-			"noGlobalRole": "Không có vai trò",
-			"globalRoleHint": "Quyết định những gì agent này có thể làm ở phạm vi toàn cục (ví dụ: từ trang chat trang chủ/quản trị). Không bắt buộc — vai trò được cấp sau, khi được mời vào một dự án, là riêng biệt.",
-			"acpProviderLabel": "Nhà cung cấp ACP",
-			"providerLabel": "Nhà cung cấp",
-			"modelLabel": "Mô hình",
-			"customOption": "Tùy chỉnh…",
-			"apiKeyLabel": "Khóa API",
-			"apiKeyKeepPlaceholder": "Để trống để giữ khóa hiện tại",
-			"baseUrlLabel": "URL cơ sở",
-			"systemPromptLabel": "System Prompt",
-			"optional": "tùy chọn",
-			"cancel": "Hủy",
-			"save": "Lưu thay đổi",
-			"updateFailed": "Cập nhật agent thất bại"
-		},
-		"deleteDialog": {
-			"title": "Xóa agent?",
-			"confirmPrefix": "Thao tác này sẽ xóa vĩnh viễn",
-			"confirmSuffix": ", xóa agent khỏi mọi dự án mà agent đã được mời tham gia. Không thể hoàn tác thao tác này.",
-			"cancel": "Hủy",
-			"delete": "Xóa",
-			"deleteFailed": "Xóa agent thất bại"
 		}
 	}
 }
