+++ apps/web/src/routes/_authenticated/projects/$projectId/agents/$agentId/index.tsx
@@ -1,69 +1,12 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
+import { AgentDetailView } from "@/components/projects/agents/agent-detail";
 import {
-	Activity as ActivityIcon,
-	Bot,
-	Check,
-	Code2,
-	KeyRound,
-	Loader2,
-	Plus,
-	Save,
-	Server,
-	Trash2,
-	Wand2,
-} from "lucide-react";
-import { useEffect, useState } from "react";
-import { useTranslation } from "react-i18next";
-import { AcpBridgeSetup } from "@/components/projects/agents/acp-bridge-setup";
-import { AgentActivityTab } from "@/components/projects/agents/agent-activity-tab";
-import { Avatar, AvatarFallback } from "@/components/ui/avatar";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import {
-	Dialog,
-	DialogContent,
-	DialogDescription,
-	DialogFooter,
-	DialogHeader,
-	DialogTitle,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectSeparator,
-	SelectTrigger,
-	SelectValue,
-} from "@/components/ui/select";
-import { Separator } from "@/components/ui/separator";
-import { Skeleton } from "@/components/ui/skeleton";
-import { Switch } from "@/components/ui/switch";
-import { Textarea } from "@/components/ui/textarea";
-import { useProjectPermissions } from "@/hooks/use-project-permissions";
-import {
-	type ACPProvider,
-	type Agent,
-	type AgentMCPServer,
-	type AgentSkill,
-	addEnvVar,
-	addMCPServer,
-	addSkill,
 	agentEnvVarsQueryOptions,
 	agentMCPServersQueryOptions,
 	agentQueryOptions,
 	agentSkillsQueryOptions,
-	deleteEnvVar,
-	deleteMCPServer,
-	deleteSkill,
 	llmModelsQueryOptions,
-	updateAgent,
-	updateMCPServer,
-	updateSkill,
 } from "@/lib/agent-api";
-import { splitShellCommand } from "@/lib/shell-command";
 
 export const Route = createFileRoute(
 	"/_authenticated/projects/$projectId/agents/$agentId/",
@@ -82,1329 +25,10 @@ export const Route = createFileRoute(
 			queryClient.ensureQueryData(llmModelsQueryOptions),
 		]);
 	},
-	component: AgentDetailPage,
+	component: ProjectAgentDetailPage,
 });
 
-type Tab = "overview" | "mcp-servers" | "skills" | "env-vars" | "activity";
-
-const CUSTOM = "__custom__";
-
-// ── Overview Tab ──────────────────────────────────────────────────────────────
-
-function OverviewTab({
-	agent,
-	projectId,
-	canWrite,
-}: {
-	agent: Agent;
-	projectId: string;
-	canWrite: boolean;
-}) {
-	const { t } = useTranslation("projects");
-	const qc = useQueryClient();
-	const { data: llmModels = {} } = useQuery(llmModelsQueryOptions);
-
-	const providers = Object.keys(llmModels);
-
-	// Provider select: if agent's provider is known, use it directly; otherwise custom mode
-	const knownProvider =
-		providers.length > 0 && providers.includes(agent.llm_provider);
-	const [providerSelect, setProviderSelect] = useState(
-		knownProvider
-			? agent.llm_provider
-			: agent.llm_provider
-				? CUSTOM
-				: "anthropic",
-	);
-	const [customProvider, setCustomProvider] = useState(
-		knownProvider ? "" : agent.llm_provider,
-	);
-
-	// Model select: check against the provider's model list once loaded
-	const initialModels = llmModels[agent.llm_provider]?.models ?? [];
-	const knownModel = initialModels.includes(agent.llm_model);
-	const [modelSelect, setModelSelect] = useState(
-		knownModel ? agent.llm_model : agent.llm_model ? CUSTOM : "",
-	);
-	const [customModel, setCustomModel] = useState(
-		knownModel ? "" : agent.llm_model,
-	);
-
-	const [name, setName] = useState(agent.name);
-	const [llmApiKey, setLlmApiKey] = useState("");
-	const [llmBaseUrl, setLlmBaseUrl] = useState(agent.llm_base_url ?? "");
-	const [systemPrompt, setSystemPrompt] = useState(agent.system_prompt);
-	const [committerName, setCommitterName] = useState(agent.git_committer_name);
-	const [committerEmail, setCommitterEmail] = useState(
-		agent.git_committer_email,
-	);
-	const [acpProviderSelect, setAcpProviderSelect] = useState<ACPProvider>(
-		agent.acp_provider ?? "claude-code",
-	);
-	const [acpCommand, setAcpCommand] = useState(
-		(agent.acp_command ?? []).join(" "),
-	);
-
-	// Derived final values sent to the API
-	const llmProvider =
-		providerSelect === CUSTOM ? customProvider.trim() : providerSelect;
-	const llmModel = modelSelect === CUSTOM ? customModel.trim() : modelSelect;
-	const acpCommandParts = splitShellCommand(acpCommand);
-	const isAcp = agent.agent_type === "acp";
-
-	const handleProviderChange = (v: string | null) => {
-		if (!v) return;
-		setProviderSelect(v);
-		if (v !== CUSTOM) {
-			const info = llmModels[v];
-			setLlmBaseUrl(info?.base_url ?? "");
-			const firstModel = info?.models?.[0] ?? "";
-			setModelSelect(firstModel || CUSTOM);
-			if (!firstModel) setCustomModel("");
-		} else {
-			setModelSelect(CUSTOM);
-			setCustomModel("");
-		}
-	};
-
-	const availableModels: string[] =
-		providerSelect !== CUSTOM ? (llmModels[providerSelect]?.models ?? []) : [];
-
-	const isDirty =
-		name !== agent.name ||
-		(isAcp
-			? acpProviderSelect !== (agent.acp_provider ?? "claude-code") ||
-				acpCommandParts.join(" ") !== (agent.acp_command ?? []).join(" ")
-			: llmProvider !== agent.llm_provider ||
-				llmModel !== agent.llm_model ||
-				llmApiKey !== "" ||
-				llmBaseUrl !== (agent.llm_base_url ?? "") ||
-				systemPrompt !== agent.system_prompt ||
-				committerName !== agent.git_committer_name ||
-				committerEmail !== agent.git_committer_email);
-
-	const saveMutation = useMutation({
-		mutationFn: () =>
-			updateAgent(projectId, agent.id, {
-				name: name.trim(),
-				...(isAcp
-					? {
-							acp_provider: acpProviderSelect,
-							...(acpProviderSelect === "custom"
-								? { acp_command: acpCommandParts }
-								: {}),
-						}
-					: {
-							llm_provider: llmProvider,
-							llm_model: llmModel,
-							...(llmApiKey ? { llm_api_key: llmApiKey } : {}),
-							llm_base_url: llmBaseUrl,
-							system_prompt: systemPrompt,
-							git_committer_name: committerName.trim(),
-							git_committer_email: committerEmail.trim(),
-						}),
-			}),
-		onSuccess: (updated) => {
-			qc.setQueryData(["projects", projectId, "agents", agent.id], updated);
-			setLlmApiKey("");
-		},
-	});
-
-	const canSave =
-		isDirty &&
-		(isAcp
-			? !!acpProviderSelect &&
-				(acpProviderSelect !== "custom" || acpCommandParts.length > 0)
-			: !!llmProvider && !!llmModel && !!llmBaseUrl.trim()) &&
-		!saveMutation.isPending;
-
-	return (
-		<div className="space-y-6 max-w-2xl">
-			<div className="space-y-1.5">
-				<Label>{t("agents.detail.overview.nameLabel")}</Label>
-				<Input
-					value={name}
-					onChange={(e) => setName(e.target.value)}
-					disabled={!canWrite}
-				/>
-			</div>
-
-			<Separator />
-
-			{!isAcp && (
-				<div>
-					<p className="text-sm font-medium mb-3">
-						{t("agents.detail.overview.llmConfiguration")}
-					</p>
-					<div className="grid grid-cols-2 gap-3">
-						<div className="space-y-1.5">
-							<Label>{t("agents.detail.overview.providerLabel")}</Label>
-							<Select
-								value={providerSelect}
-								onValueChange={handleProviderChange}
-								disabled={!canWrite}
-							>
-								<SelectTrigger>
-									<SelectValue />
-								</SelectTrigger>
-								<SelectContent>
-									{providers.map((p) => (
-										<SelectItem key={p} value={p}>
-											{p}
-										</SelectItem>
-									))}
-									<SelectSeparator />
-									<SelectItem value={CUSTOM}>
-										{t("agents.detail.overview.customOption")}
-									</SelectItem>
-								</SelectContent>
-							</Select>
-							{providerSelect === CUSTOM && (
-								<Input
-									placeholder="my-provider"
-									value={customProvider}
-									onChange={(e) => setCustomProvider(e.target.value)}
-									disabled={!canWrite}
-								/>
-							)}
-						</div>
-						<div className="space-y-1.5">
-							<Label>{t("agents.detail.overview.modelLabel")}</Label>
-							{providerSelect === CUSTOM ? (
-								<Input
-									placeholder="my-model-name"
-									value={customModel}
-									onChange={(e) => setCustomModel(e.target.value)}
-									disabled={!canWrite}
-								/>
-							) : (
-								<>
-									<Select
-										value={modelSelect}
-										onValueChange={(v) => v && setModelSelect(v)}
-										disabled={!canWrite}
-									>
-										<SelectTrigger>
-											<SelectValue />
-										</SelectTrigger>
-										<SelectContent>
-											{availableModels.map((m) => (
-												<SelectItem key={m} value={m}>
-													{m}
-												</SelectItem>
-											))}
-											<SelectSeparator />
-											<SelectItem value={CUSTOM}>
-												{t("agents.detail.overview.customOption")}
-											</SelectItem>
-										</SelectContent>
-									</Select>
-									{modelSelect === CUSTOM && (
-										<Input
-											placeholder="my-model-name"
-											value={customModel}
-											onChange={(e) => setCustomModel(e.target.value)}
-											disabled={!canWrite}
-										/>
-									)}
-								</>
-							)}
-						</div>
-					</div>
-					<div className="space-y-1.5 mt-3">
-						<Label>
-							{t("agents.detail.overview.apiKeyUpdateLabel")}{" "}
-							<span className="text-muted-foreground font-normal text-xs">
-								{t("agents.detail.overview.apiKeyUpdateHint")}
-							</span>
-						</Label>
-						<Input
-							type="password"
-							placeholder="sk-ant-…"
-							value={llmApiKey}
-							onChange={(e) => setLlmApiKey(e.target.value)}
-							disabled={!canWrite}
-						/>
-					</div>
-					<div className="space-y-1.5 mt-3">
-						<Label>
-							{t("agents.detail.overview.baseUrlLabel")}{" "}
-							<span className="text-destructive">*</span>
-						</Label>
-						<Input
-							placeholder="https://api.openai.com/v1"
-							value={llmBaseUrl}
-							onChange={(e) => setLlmBaseUrl(e.target.value)}
-							disabled={!canWrite}
-						/>
-					</div>
-				</div>
-			)}
-
-			{isAcp && (
-				<div>
-					<p className="text-sm font-medium mb-3">
-						{t("agents.detail.overview.acpConfiguration")}
-					</p>
-					<div className="space-y-1.5">
-						<Label>{t("agents.detail.overview.acpProviderLabel")}</Label>
-						<Select
-							value={acpProviderSelect}
-							onValueChange={(v) => v && setAcpProviderSelect(v as ACPProvider)}
-							disabled={!canWrite}
-						>
-							<SelectTrigger>
-								<SelectValue />
-							</SelectTrigger>
-							<SelectContent>
-								<SelectItem value="claude-code">
-									{t("agents.detail.overview.acpProviderClaudeCode")}
-								</SelectItem>
-								<SelectItem value="codex">
-									{t("agents.detail.overview.acpProviderCodex")}
-								</SelectItem>
-								<SelectItem value="gemini-cli">
-									{t("agents.detail.overview.acpProviderGeminiCli")}
-								</SelectItem>
-								<SelectSeparator />
-								<SelectItem value="custom">
-									{t("agents.detail.overview.acpProviderCustom")}
-								</SelectItem>
-							</SelectContent>
-						</Select>
-					</div>
-					{acpProviderSelect === "custom" && (
-						<div className="space-y-1.5 mt-3">
-							<Label>{t("agents.detail.overview.acpCommandLabel")}</Label>
-							<Input
-								placeholder="npx -y my-acp-server"
-								value={acpCommand}
-								onChange={(e) => setAcpCommand(e.target.value)}
-								disabled={!canWrite}
-							/>
-							<p className="text-xs text-muted-foreground">
-								{t("agents.detail.overview.acpCommandHint")}
-							</p>
-						</div>
-					)}
-					<p className="text-xs text-muted-foreground rounded-md bg-muted/40 px-3 py-2 mt-3">
-						{t("agents.detail.overview.acpGuidanceBanner")}
-					</p>
-				</div>
-			)}
-
-			<Separator />
-
-			{isAcp && (
-				<>
-					<LocalBridgePanel
-						projectId={projectId}
-						agentId={agent.id}
-						acpProvider={agent.acp_provider ?? "claude-code"}
-						hasToken={agent.has_acp_bridge_token}
-						canWrite={canWrite}
-						onTokenGenerated={() =>
-							qc.setQueryData(
-								["projects", projectId, "agents", agent.id],
-								(old: Agent | undefined) =>
-									old ? { ...old, has_acp_bridge_token: true } : old,
-							)
-						}
-					/>
-					<Separator />
-				</>
-			)}
-
-			{/* System prompt and git committer identity are LLM-only — an ACP
-			    agent's local CLI owns its own prompt and git identity. */}
-			{!isAcp && (
-				<>
-					<div className="space-y-1.5">
-						<Label>{t("agents.detail.overview.systemPromptLabel")}</Label>
-						<Textarea
-							value={systemPrompt}
-							onChange={(e) => setSystemPrompt(e.target.value)}
-							rows={5}
-							disabled={!canWrite}
-							className="font-mono text-xs"
-						/>
-					</div>
-
-					<Separator />
-
-					<div>
-						<p className="text-sm font-medium mb-1">
-							{t("agents.detail.overview.gitCommitterIdentity")}
-						</p>
-						<p className="text-xs text-muted-foreground mb-3">
-							{t("agents.detail.overview.gitCommitterHint")}
-						</p>
-						<div className="grid grid-cols-2 gap-3">
-							<div className="space-y-1.5">
-								<Label>{t("agents.detail.overview.committerNameLabel")}</Label>
-								<Input
-									value={committerName}
-									onChange={(e) => setCommitterName(e.target.value)}
-									disabled={!canWrite}
-									placeholder="paca-agent"
-								/>
-							</div>
-							<div className="space-y-1.5">
-								<Label>{t("agents.detail.overview.committerEmailLabel")}</Label>
-								<Input
-									type="email"
-									value={committerEmail}
-									onChange={(e) => setCommitterEmail(e.target.value)}
-									disabled={!canWrite}
-									placeholder="paca-agent@users.noreply.github.com"
-								/>
-							</div>
-						</div>
-					</div>
-				</>
-			)}
-
-			{canWrite && (
-				<div className="flex items-center gap-3 pt-2">
-					<Button onClick={() => saveMutation.mutate()} disabled={!canSave}>
-						{saveMutation.isPending ? (
-							<Loader2 className="size-4 mr-2 animate-spin" />
-						) : (
-							<Save className="size-4 mr-2" />
-						)}
-						{t("agents.detail.overview.saveChanges")}
-					</Button>
-					{saveMutation.isSuccess && (
-						<span className="flex items-center gap-1 text-xs text-emerald-600">
-							<Check className="size-3" />
-							{t("agents.detail.overview.saved")}
-						</span>
-					)}
-					{saveMutation.isError && (
-						<span className="text-xs text-destructive">
-							{t("agents.detail.overview.saveFailed")}
-						</span>
-					)}
-				</div>
-			)}
-		</div>
-	);
-}
-
-// ── Local Bridge Panel (ACP agents, embedded in Overview) ───────────────────────
-
-function LocalBridgePanel({
-	projectId,
-	agentId,
-	acpProvider,
-	hasToken,
-	canWrite,
-	onTokenGenerated,
-}: {
-	projectId: string;
-	agentId: string;
-	acpProvider: ACPProvider;
-	hasToken: boolean;
-	canWrite: boolean;
-	onTokenGenerated: () => void;
-}) {
-	const { t } = useTranslation("projects");
-	return (
-		<div>
-			<p className="text-sm font-medium mb-1">
-				{t("agents.acpSetup.panelTitle")}
-			</p>
-			<p className="text-xs text-muted-foreground mb-3">
-				{t("agents.acpSetup.panelDescription")}
-			</p>
-			<AcpBridgeSetup
-				projectId={projectId}
-				agentId={agentId}
-				acpProvider={acpProvider}
-				hasToken={hasToken}
-				canWrite={canWrite}
-				onTokenGenerated={onTokenGenerated}
-			/>
-		</div>
-	);
-}
-
-// ── MCP Servers Tab ───────────────────────────────────────────────────────────
-
-function AddMCPServerDialog({
-	projectId,
-	agentId,
-	open,
-	onOpenChange,
-}: {
-	projectId: string;
-	agentId: string;
-	open: boolean;
-	onOpenChange: (open: boolean) => void;
-}) {
-	const { t } = useTranslation("projects");
-	const qc = useQueryClient();
-	const [serverName, setServerName] = useState("");
-	const [transport, setTransport] = useState<"stdio" | "sse" | "http">("stdio");
-	const [command, setCommand] = useState("");
-	const [args, setArgs] = useState("");
-	const [url, setUrl] = useState("");
-
-	const addMutation = useMutation({
-		mutationFn: () =>
-			addMCPServer(projectId, agentId, {
-				server_name: serverName.trim(),
-				transport,
-				command: transport === "stdio" ? command.trim() || null : null,
-				args:
-					transport === "stdio"
-						? args
-								.split(/\s+/)
-								.map((a) => a.trim())
-								.filter(Boolean)
-						: [],
-				url: transport !== "stdio" ? url.trim() || null : null,
-			}),
-		onSuccess: () => {
-			qc.invalidateQueries({
-				queryKey: ["projects", projectId, "agents", agentId, "mcp-servers"],
-			});
-			onOpenChange(false);
-			setServerName("");
-			setCommand("");
-			setArgs("");
-			setUrl("");
-		},
-	});
-
-	return (
-		<Dialog open={open} onOpenChange={onOpenChange}>
-			<DialogContent className="max-w-md">
-				<DialogHeader>
-					<DialogTitle className="flex items-center gap-2">
-						<Server className="size-4 text-primary" />
-						{t("agents.detail.mcp.addDialog.title")}
-					</DialogTitle>
-					<DialogDescription>
-						{t("agents.detail.mcp.addDialog.description")}
-					</DialogDescription>
-				</DialogHeader>
-				<div className="space-y-4 py-2">
-					<div className="space-y-1.5">
-						<Label>{t("agents.detail.mcp.addDialog.serverNameLabel")}</Label>
-						<Input
-							placeholder="filesystem"
-							value={serverName}
-							onChange={(e) => setServerName(e.target.value)}
-						/>
-					</div>
-					<div className="space-y-1.5">
-						<Label>{t("agents.detail.mcp.addDialog.transportLabel")}</Label>
-						<Select
-							value={transport}
-							onValueChange={(v) => setTransport(v as typeof transport)}
-						>
-							<SelectTrigger>
-								<SelectValue />
-							</SelectTrigger>
-							<SelectContent>
-								<SelectItem value="stdio">stdio</SelectItem>
-								<SelectItem value="sse">SSE</SelectItem>
-								<SelectItem value="http">HTTP</SelectItem>
-							</SelectContent>
-						</Select>
-					</div>
-					{transport === "stdio" ? (
-						<>
-							<div className="space-y-1.5">
-								<Label>{t("agents.detail.mcp.addDialog.commandLabel")}</Label>
-								<Input
-									placeholder="npx"
-									value={command}
-									onChange={(e) => setCommand(e.target.value)}
-								/>
-							</div>
-							<div className="space-y-1.5">
-								<Label>
-									{t("agents.detail.mcp.addDialog.argsLabel")}{" "}
-									<span className="text-muted-foreground font-normal text-xs">
-										{t("agents.detail.mcp.addDialog.argsHint")}
-									</span>
-								</Label>
-								<Input
-									placeholder="-y @modelcontextprotocol/server-filesystem /tmp"
-									value={args}
-									onChange={(e) => setArgs(e.target.value)}
-								/>
-							</div>
-						</>
-					) : (
-						<div className="space-y-1.5">
-							<Label>{t("agents.detail.mcp.addDialog.urlLabel")}</Label>
-							<Input
-								placeholder="https://mcp.example.com/sse"
-								value={url}
-								onChange={(e) => setUrl(e.target.value)}
-							/>
-						</div>
-					)}
-				</div>
-				<DialogFooter>
-					<Button variant="outline" onClick={() => onOpenChange(false)}>
-						{t("agents.detail.mcp.addDialog.cancel")}
-					</Button>
-					<Button
-						onClick={() => addMutation.mutate()}
-						disabled={!serverName.trim() || addMutation.isPending}
-					>
-						{addMutation.isPending ? (
-							<Loader2 className="size-4 animate-spin" />
-						) : (
-							t("agents.detail.mcp.addDialog.addServer")
-						)}
-					</Button>
-				</DialogFooter>
-			</DialogContent>
-		</Dialog>
-	);
-}
-
-function MCPServersTab({
-	projectId,
-	agentId,
-	canWrite,
-}: {
-	projectId: string;
-	agentId: string;
-	canWrite: boolean;
-}) {
-	const { t } = useTranslation("projects");
-	const qc = useQueryClient();
-	const { data: servers = [] } = useQuery(
-		agentMCPServersQueryOptions(projectId, agentId),
-	);
-	const [addOpen, setAddOpen] = useState(false);
-
-	const toggleMutation = useMutation({
-		mutationFn: (s: AgentMCPServer) =>
-			updateMCPServer(projectId, agentId, s.id, {
-				is_enabled: !s.is_enabled,
-			}),
-		onSuccess: () => {
-			qc.invalidateQueries({
-				queryKey: ["projects", projectId, "agents", agentId, "mcp-servers"],
-			});
-		},
-	});
-
-	const deleteMutation = useMutation({
-		mutationFn: (id: string) => deleteMCPServer(projectId, agentId, id),
-		onSuccess: () => {
-			qc.invalidateQueries({
-				queryKey: ["projects", projectId, "agents", agentId, "mcp-servers"],
-			});
-		},
-	});
-
-	return (
-		<div className="space-y-4">
-			<div className="flex items-center justify-between">
-				<p className="text-sm text-muted-foreground">
-					{t("agents.detail.mcp.serverCount", { count: servers.length })}
-				</p>
-				{canWrite && (
-					<Button size="sm" onClick={() => setAddOpen(true)}>
-						<Plus className="size-4 mr-1.5" />
-						{t("agents.detail.mcp.addServer")}
-					</Button>
-				)}
-			</div>
-
-			{servers.length === 0 ? (
-				<div className="flex flex-col items-center justify-center gap-3 py-14 rounded-xl border border-dashed border-border">
-					<Server className="size-8 text-muted-foreground/40" />
-					<p className="text-sm text-muted-foreground">
-						{t("agents.detail.mcp.empty.title")}
-					</p>
-					{canWrite && (
-						<Button
-							size="sm"
-							variant="outline"
-							onClick={() => setAddOpen(true)}
-						>
-							<Plus className="size-3.5 mr-1" />
-							{t("agents.detail.mcp.empty.addFirstServer")}
-						</Button>
-					)}
-				</div>
-			) : (
-				<div className="space-y-2">
-					{servers.map((s) => (
-						<div
-							key={s.id}
-							className="flex items-center justify-between gap-3 rounded-lg border border-border/60 bg-card px-4 py-3"
-						>
-							<div className="flex items-center gap-3 min-w-0">
-								<Server className="size-4 text-muted-foreground shrink-0" />
-								<div className="min-w-0">
-									<p className="text-sm font-medium truncate">
-										{s.server_name}
-									</p>
-									<p className="text-xs text-muted-foreground font-mono truncate">
-										{s.transport}
-										{s.command ? ` · ${s.command}` : ""}
-										{s.url ? ` · ${s.url}` : ""}
-									</p>
-								</div>
-							</div>
-							<div className="flex items-center gap-2 shrink-0">
-								<Switch
-									checked={s.is_enabled}
-									onCheckedChange={() => canWrite && toggleMutation.mutate(s)}
-									disabled={!canWrite || toggleMutation.isPending}
-								/>
-								{canWrite && (
-									<Button
-										variant="ghost"
-										size="icon"
-										className="size-7 text-muted-foreground hover:text-destructive"
-										onClick={() => deleteMutation.mutate(s.id)}
-										disabled={deleteMutation.isPending}
-									>
-										<Trash2 className="size-3.5" />
-									</Button>
-								)}
-							</div>
-						</div>
-					))}
-				</div>
-			)}
-
-			<AddMCPServerDialog
-				projectId={projectId}
-				agentId={agentId}
-				open={addOpen}
-				onOpenChange={setAddOpen}
-			/>
-		</div>
-	);
-}
-
-// ── Skills Tab ────────────────────────────────────────────────────────────────
-
-function AddSkillDialog({
-	projectId,
-	agentId,
-	open,
-	onOpenChange,
-}: {
-	projectId: string;
-	agentId: string;
-	open: boolean;
-	onOpenChange: (open: boolean) => void;
-}) {
-	const { t } = useTranslation("projects");
-	const qc = useQueryClient();
-	const [skillName, setSkillName] = useState("");
-	const [source, setSource] = useState<"inline" | "marketplace" | "github_url">(
-		"inline",
-	);
-	const [skillContent, setSkillContent] = useState("");
-	const [sourceUrl, setSourceUrl] = useState("");
-
-	const addMutation = useMutation({
-		mutationFn: () =>
-			addSkill(projectId, agentId, {
-				skill_name: skillName.trim(),
-				skill_source: source,
-				skill_content: source === "inline" ? skillContent : undefined,
-				source_url: source !== "inline" ? sourceUrl.trim() : null,
-			}),
-		onSuccess: () => {
-			qc.invalidateQueries({
-				queryKey: ["projects", projectId, "agents", agentId, "skills"],
-			});
-			onOpenChange(false);
-			setSkillName("");
-			setSkillContent("");
-			setSourceUrl("");
-		},
-	});
-
-	return (
-		<Dialog open={open} onOpenChange={onOpenChange}>
-			<DialogContent className="max-w-md">
-				<DialogHeader>
-					<DialogTitle className="flex items-center gap-2">
-						<Wand2 className="size-4 text-primary" />
-						{t("agents.detail.skills.addDialog.title")}
-					</DialogTitle>
-					<DialogDescription>
-						{t("agents.detail.skills.addDialog.description")}
-					</DialogDescription>
-				</DialogHeader>
-				<div className="space-y-4 py-2">
-					<div className="space-y-1.5">
-						<Label>{t("agents.detail.skills.addDialog.skillNameLabel")}</Label>
-						<Input
-							placeholder="code-reviewer"
-							value={skillName}
-							onChange={(e) => setSkillName(e.target.value)}
-						/>
-					</div>
-					<div className="space-y-1.5">
-						<Label>{t("agents.detail.skills.addDialog.sourceLabel")}</Label>
-						<Select
-							value={source}
-							onValueChange={(v) => setSource(v as typeof source)}
-						>
-							<SelectTrigger>
-								<SelectValue />
-							</SelectTrigger>
-							<SelectContent>
-								<SelectItem value="inline">
-									{t("agents.detail.skills.addDialog.sourceInline")}
-								</SelectItem>
-								<SelectItem value="marketplace">
-									{t("agents.detail.skills.addDialog.sourceMarketplace")}
-								</SelectItem>
-								<SelectItem value="github_url">
-									{t("agents.detail.skills.addDialog.sourceGithubUrl")}
-								</SelectItem>
-							</SelectContent>
-						</Select>
-					</div>
-					{source === "inline" ? (
-						<div className="space-y-1.5">
-							<Label>
-								{t("agents.detail.skills.addDialog.skillContentLabel")}
-							</Label>
-							<Textarea
-								placeholder="# Code Reviewer&#10;&#10;You review pull requests for security, performance…"
-								value={skillContent}
-								onChange={(e) => setSkillContent(e.target.value)}
-								rows={5}
-								className="font-mono text-xs"
-							/>
-						</div>
-					) : (
-						<div className="space-y-1.5">
-							<Label>{t("agents.detail.skills.addDialog.urlLabel")}</Label>
-							<Input
-								placeholder={
-									source === "marketplace"
-										? "paca/code-reviewer@1.0.0"
-										: "https://github.com/org/skills/blob/main/SKILL.md"
-								}
-								value={sourceUrl}
-								onChange={(e) => setSourceUrl(e.target.value)}
-							/>
-						</div>
-					)}
-				</div>
-				<DialogFooter>
-					<Button variant="outline" onClick={() => onOpenChange(false)}>
-						{t("agents.detail.skills.addDialog.cancel")}
-					</Button>
-					<Button
-						onClick={() => addMutation.mutate()}
-						disabled={!skillName.trim() || addMutation.isPending}
-					>
-						{addMutation.isPending ? (
-							<Loader2 className="size-4 animate-spin" />
-						) : (
-							t("agents.detail.skills.addDialog.addSkill")
-						)}
-					</Button>
-				</DialogFooter>
-			</DialogContent>
-		</Dialog>
-	);
-}
-
-function SkillsTab({
-	projectId,
-	agentId,
-	canWrite,
-}: {
-	projectId: string;
-	agentId: string;
-	canWrite: boolean;
-}) {
-	const { t } = useTranslation("projects");
-	const qc = useQueryClient();
-	const { data: skills = [] } = useQuery(
-		agentSkillsQueryOptions(projectId, agentId),
-	);
-	const [addOpen, setAddOpen] = useState(false);
-
-	const toggleMutation = useMutation({
-		mutationFn: (s: AgentSkill) =>
-			updateSkill(projectId, agentId, s.id, { is_enabled: !s.is_enabled }),
-		onSuccess: () => {
-			qc.invalidateQueries({
-				queryKey: ["projects", projectId, "agents", agentId, "skills"],
-			});
-		},
-	});
-
-	const deleteMutation = useMutation({
-		mutationFn: (id: string) => deleteSkill(projectId, agentId, id),
-		onSuccess: () => {
-			qc.invalidateQueries({
-				queryKey: ["projects", projectId, "agents", agentId, "skills"],
-			});
-		},
-	});
-
-	return (
-		<div className="space-y-4">
-			<div className="flex items-center justify-between">
-				<p className="text-sm text-muted-foreground">
-					{t("agents.detail.skills.skillCount", { count: skills.length })}
-				</p>
-				{canWrite && (
-					<Button size="sm" onClick={() => setAddOpen(true)}>
-						<Plus className="size-4 mr-1.5" />
-						{t("agents.detail.skills.addSkill")}
-					</Button>
-				)}
-			</div>
-
-			{skills.length === 0 ? (
-				<div className="flex flex-col items-center justify-center gap-3 py-14 rounded-xl border border-dashed border-border">
-					<Wand2 className="size-8 text-muted-foreground/40" />
-					<p className="text-sm text-muted-foreground">
-						{t("agents.detail.skills.empty.title")}
-					</p>
-					{canWrite && (
-						<Button
-							size="sm"
-							variant="outline"
-							onClick={() => setAddOpen(true)}
-						>
-							<Plus className="size-3.5 mr-1" />
-							{t("agents.detail.skills.empty.addFirstSkill")}
-						</Button>
-					)}
-				</div>
-			) : (
-				<div className="space-y-2">
-					{skills.map((s) => (
-						<div
-							key={s.id}
-							className="flex items-center justify-between gap-3 rounded-lg border border-border/60 bg-card px-4 py-3"
-						>
-							<div className="flex items-center gap-3 min-w-0">
-								<Code2 className="size-4 text-muted-foreground shrink-0" />
-								<div className="min-w-0">
-									<p className="text-sm font-medium truncate">{s.skill_name}</p>
-									<p className="text-xs text-muted-foreground">
-										{s.skill_source}
-										{s.source_url ? ` · ${s.source_url}` : ""}
-									</p>
-								</div>
-							</div>
-							<div className="flex items-center gap-2 shrink-0">
-								<Switch
-									checked={s.is_enabled}
-									onCheckedChange={() => canWrite && toggleMutation.mutate(s)}
-									disabled={!canWrite || toggleMutation.isPending}
-								/>
-								{canWrite && (
-									<Button
-										variant="ghost"
-										size="icon"
-										className="size-7 text-muted-foreground hover:text-destructive"
-										onClick={() => deleteMutation.mutate(s.id)}
-										disabled={deleteMutation.isPending}
-									>
-										<Trash2 className="size-3.5" />
-									</Button>
-								)}
-							</div>
-						</div>
-					))}
-				</div>
-			)}
-
-			<AddSkillDialog
-				projectId={projectId}
-				agentId={agentId}
-				open={addOpen}
-				onOpenChange={setAddOpen}
-			/>
-		</div>
-	);
-}
-
-// ── Environment Variables Tab ────────────────────────────────────────────────
-
-const ENV_VAR_KEY_PATTERN = /^[A-Za-z_][A-Za-z0-9_]*$/;
-
-function AddEnvVarDialog({
-	projectId,
-	agentId,
-	open,
-	onOpenChange,
-}: {
-	projectId: string;
-	agentId: string;
-	open: boolean;
-	onOpenChange: (open: boolean) => void;
-}) {
-	const { t } = useTranslation("projects");
-	const qc = useQueryClient();
-	const [key, setKey] = useState("");
-	const [value, setValue] = useState("");
-
-	const isKeyValid = ENV_VAR_KEY_PATTERN.test(key.trim());
-
-	const addMutation = useMutation({
-		mutationFn: () => addEnvVar(projectId, agentId, { key: key.trim(), value }),
-		onSuccess: () => {
-			qc.invalidateQueries({
-				queryKey: ["projects", projectId, "agents", agentId, "env-vars"],
-			});
-			onOpenChange(false);
-			setKey("");
-			setValue("");
-		},
-	});
-
-	return (
-		<Dialog open={open} onOpenChange={onOpenChange}>
-			<DialogContent className="max-w-md">
-				<DialogHeader>
-					<DialogTitle className="flex items-center gap-2">
-						<KeyRound className="size-4 text-primary" />
-						{t("agents.detail.envVars.addDialog.title")}
-					</DialogTitle>
-					<DialogDescription>
-						{t("agents.detail.envVars.addDialog.description")}
-					</DialogDescription>
-				</DialogHeader>
-				<div className="space-y-4 py-2">
-					<div className="space-y-1.5">
-						<Label>{t("agents.detail.envVars.addDialog.keyLabel")}</Label>
-						<Input
-							placeholder={t("agents.detail.envVars.addDialog.keyPlaceholder")}
-							className="font-mono"
-							value={key}
-							onChange={(e) => setKey(e.target.value)}
-						/>
-						<p className="text-xs text-muted-foreground">
-							{t("agents.detail.envVars.addDialog.keyHint")}
-						</p>
-					</div>
-					<div className="space-y-1.5">
-						<Label>{t("agents.detail.envVars.addDialog.valueLabel")}</Label>
-						<Input
-							type="password"
-							autoComplete="off"
-							className="font-mono"
-							value={value}
-							onChange={(e) => setValue(e.target.value)}
-						/>
-					</div>
-				</div>
-				<DialogFooter>
-					<Button variant="outline" onClick={() => onOpenChange(false)}>
-						{t("agents.detail.envVars.addDialog.cancel")}
-					</Button>
-					<Button
-						onClick={() => addMutation.mutate()}
-						disabled={!isKeyValid || !value || addMutation.isPending}
-					>
-						{addMutation.isPending ? (
-							<Loader2 className="size-4 animate-spin" />
-						) : (
-							t("agents.detail.envVars.addDialog.addVariable")
-						)}
-					</Button>
-				</DialogFooter>
-			</DialogContent>
-		</Dialog>
-	);
-}
-
-function EnvVarsTab({
-	projectId,
-	agentId,
-	canWrite,
-}: {
-	projectId: string;
-	agentId: string;
-	canWrite: boolean;
-}) {
-	const { t } = useTranslation("projects");
-	const qc = useQueryClient();
-	const { data: envVars = [] } = useQuery(
-		agentEnvVarsQueryOptions(projectId, agentId),
-	);
-	const [addOpen, setAddOpen] = useState(false);
-
-	const deleteMutation = useMutation({
-		mutationFn: (id: string) => deleteEnvVar(projectId, agentId, id),
-		onSuccess: () => {
-			qc.invalidateQueries({
-				queryKey: ["projects", projectId, "agents", agentId, "env-vars"],
-			});
-		},
-	});
-
-	return (
-		<div className="space-y-4">
-			<div className="flex items-center justify-between">
-				<p className="text-sm text-muted-foreground">
-					{t("agents.detail.envVars.count", { count: envVars.length })}
-				</p>
-				{canWrite && (
-					<Button size="sm" onClick={() => setAddOpen(true)}>
-						<Plus className="size-4 mr-1.5" />
-						{t("agents.detail.envVars.addVariable")}
-					</Button>
-				)}
-			</div>
-
-			{envVars.length === 0 ? (
-				<div className="flex flex-col items-center justify-center gap-3 py-14 rounded-xl border border-dashed border-border">
-					<KeyRound className="size-8 text-muted-foreground/40" />
-					<p className="text-sm text-muted-foreground">
-						{t("agents.detail.envVars.empty.title")}
-					</p>
-					{canWrite && (
-						<Button
-							size="sm"
-							variant="outline"
-							onClick={() => setAddOpen(true)}
-						>
-							<Plus className="size-3.5 mr-1" />
-							{t("agents.detail.envVars.empty.addFirst")}
-						</Button>
-					)}
-				</div>
-			) : (
-				<div className="space-y-2">
-					{envVars.map((v) => (
-						<div
-							key={v.id}
-							className="flex items-center justify-between gap-3 rounded-lg border border-border/60 bg-card px-4 py-3"
-						>
-							<div className="flex items-center gap-3 min-w-0">
-								<KeyRound className="size-4 text-muted-foreground shrink-0" />
-								<div className="min-w-0">
-									<p className="text-sm font-medium font-mono truncate">
-										{v.key}
-									</p>
-									<p className="text-xs text-muted-foreground font-mono truncate">
-										{v.value}
-									</p>
-								</div>
-							</div>
-							{canWrite && (
-								<Button
-									variant="ghost"
-									size="icon"
-									className="size-7 text-muted-foreground hover:text-destructive shrink-0"
-									onClick={() => deleteMutation.mutate(v.id)}
-									disabled={deleteMutation.isPending}
-								>
-									<Trash2 className="size-3.5" />
-								</Button>
-							)}
-						</div>
-					))}
-				</div>
-			)}
-
-			<AddEnvVarDialog
-				projectId={projectId}
-				agentId={agentId}
-				open={addOpen}
-				onOpenChange={setAddOpen}
-			/>
-		</div>
-	);
-}
-
-// ── Page ──────────────────────────────────────────────────────────────────────
-
-const TABS = [
-	{ id: "overview", labelKey: "agents.detail.tabs.overview", icon: Bot },
-	{
-		id: "mcp-servers",
-		labelKey: "agents.detail.tabs.mcpServers",
-		icon: Server,
-	},
-	{ id: "skills", labelKey: "agents.detail.tabs.skills", icon: Wand2 },
-	{
-		id: "env-vars",
-		labelKey: "agents.detail.tabs.envVars",
-		icon: KeyRound,
-	},
-	{
-		id: "activity",
-		labelKey: "agents.detail.tabs.activity",
-		icon: ActivityIcon,
-	},
-] as const satisfies {
-	id: Tab;
-	labelKey: string;
-	icon: React.ComponentType<{ className?: string }>;
-}[];
-
-function AgentDetailPage() {
-	const { t } = useTranslation("projects");
+function ProjectAgentDetailPage() {
 	const { projectId, agentId } = Route.useParams();
-	const { hasProjectPermission } = useProjectPermissions(projectId);
-	const canWrite = hasProjectPermission("agents.write");
-
-	const { data: agent } = useQuery(agentQueryOptions(projectId, agentId));
-	const [activeTab, setActiveTab] = useState<Tab>(() => {
-		const hash = window.location.hash.slice(1);
-		if (hash && TABS.map((t) => t.id).includes(hash as Tab)) {
-			return hash as Tab;
-		}
-		return "overview";
-	});
-
-	// Sync tab when hash changes (e.g., back button)
-	useEffect(() => {
-		const handleHashChange = () => {
-			const hash = window.location.hash.slice(1);
-			if (hash && TABS.map((t) => t.id).includes(hash as Tab)) {
-				setActiveTab(hash as Tab);
-			}
-		};
-		window.addEventListener("hashchange", handleHashChange);
-		return () => window.removeEventListener("hashchange", handleHashChange);
-	}, []);
-
-	const handleTabChange = (tab: Tab) => {
-		setActiveTab(tab);
-		const url = new URL(window.location.href);
-		url.hash = tab;
-		window.history.pushState(null, "", url);
-	};
-
-	if (!agent) {
-		return (
-			<div className="flex flex-col gap-4 p-6">
-				<Skeleton className="h-16 w-full rounded-xl" />
-				<Skeleton className="h-64 w-full rounded-xl" />
-			</div>
-		);
-	}
-
-	const initials = agent.name
-		.split(" ")
-		.map((w) => w[0])
-		.join("")
-		.toUpperCase()
-		.slice(0, 2);
-
-	// ACP agents run entirely in the user's own local environment (see the
-	// Local Bridge panel on the Overview tab) — Paca never forwards tools,
-	// MCP servers, skills, or environment variables into that local process,
-	// so none of those tabs apply.
-	const acpHiddenTabs: Tab[] = ["mcp-servers", "skills", "env-vars"];
-	const visibleTabs = TABS.filter(
-		(tab) => agent.agent_type !== "acp" || !acpHiddenTabs.includes(tab.id),
-	);
-
-	return (
-		<div className="flex flex-col flex-1 min-h-0">
-			{/* Agent header */}
-			<div className="border-b border-border/50 px-6 py-5 shrink-0">
-				<div className="flex items-center gap-4">
-					<Avatar className="size-12 rounded-xl bg-primary/10">
-						<AvatarFallback className="rounded-xl bg-primary/10 text-primary font-bold text-base">
-							{initials}
-						</AvatarFallback>
-					</Avatar>
-					<div>
-						<h1 className="text-lg font-semibold">{agent.name}</h1>
-						<div className="flex items-center gap-2 mt-0.5">
-							<span className="text-sm text-muted-foreground">
-								@{agent.handle}
-							</span>
-							<span className="text-muted-foreground/40">·</span>
-							<Badge variant="secondary" className="text-xs">
-								{agent.agent_type === "acp"
-									? (agent.acp_provider ?? "acp")
-									: agent.llm_provider}
-							</Badge>
-						</div>
-					</div>
-				</div>
-			</div>
-
-			{/* Tabs */}
-			<div className="border-b border-border/50 px-6 shrink-0">
-				<div className="flex items-center gap-1 -mb-px">
-					{visibleTabs.map((tab) => {
-						const Icon = tab.icon;
-						const isActive = activeTab === tab.id;
-						return (
-							<button
-								key={tab.id}
-								type="button"
-								onClick={() => handleTabChange(tab.id)}
-								className={`flex items-center gap-1.5 px-3 py-2.5 text-sm font-medium border-b-2 transition-colors ${
-									isActive
-										? "border-primary text-primary"
-										: "border-transparent text-muted-foreground hover:text-foreground"
-								}`}
-							>
-								<Icon className="size-3.5" />
-								{t(tab.labelKey)}
-							</button>
-						);
-					})}
-				</div>
-			</div>
-
-			{/* Tab content */}
-			<div
-				className={
-					activeTab === "activity"
-						? "flex flex-1 min-h-0 flex-col p-6"
-						: "flex-1 overflow-auto p-6"
-				}
-			>
-				{activeTab === "overview" && (
-					<OverviewTab
-						agent={agent}
-						projectId={projectId}
-						canWrite={canWrite}
-					/>
-				)}
-				{activeTab === "mcp-servers" && agent.agent_type !== "acp" && (
-					<MCPServersTab
-						projectId={projectId}
-						agentId={agentId}
-						canWrite={canWrite}
-					/>
-				)}
-				{activeTab === "skills" && agent.agent_type !== "acp" && (
-					<SkillsTab
-						projectId={projectId}
-						agentId={agentId}
-						canWrite={canWrite}
-					/>
-				)}
-				{activeTab === "env-vars" && agent.agent_type !== "acp" && (
-					<EnvVarsTab
-						projectId={projectId}
-						agentId={agentId}
-						canWrite={canWrite}
-					/>
-				)}
-				{activeTab === "activity" && (
-					<AgentActivityTab projectId={projectId} agentId={agentId} />
-				)}
-			</div>
-		</div>
-	);
+	return <AgentDetailView projectId={projectId} agentId={agentId} />;
 }
