+++ services/ai-agent/src/agent/builder.py
@@ -199,6 +199,7 @@ def build_mcp_config(
     db_servers: list[AgentMCPServerRow],
     agent_id: str,
     project_id: str | None,
+    actor_user_id: str | None = None,
 ) -> dict:
     """Build the MCP server configuration dict for the OpenHands SDK.
 
@@ -214,6 +215,14 @@ def build_mcp_config(
     target any project the agent has permission in, rather than being locked
     to one (see apps/mcp/src/index.ts's PACA_AGENT_ID/PACA_PROJECT_ID
     handling).
+
+    actor_user_id is the human the conversation's trigger names (e.g. a
+    global-chat message) — forwarded as PACA_ACTOR_USER_ID so tool calls like
+    "create a project for me" attribute the result to that human rather than
+    the shared agent-bot identity. Safe to trust here because this whole
+    process tree is server-side and this value came from DB-verified trigger
+    data, never from LLM-controllable tool-call arguments (see
+    services/api's parseActorUserIDHeader for the matching server-side check).
     """
     servers: dict = {}
 
@@ -245,6 +254,8 @@ def build_mcp_config(
         }
         if project_id is not None:
             env["PACA_PROJECT_ID"] = project_id
+        if actor_user_id is not None:
+            env["PACA_ACTOR_USER_ID"] = actor_user_id
         servers["paca"] = {
             "command": command,
             "args": args,
